+++ sdk/resourcemanager/appservice/armappservice/models.go
@@ -5632,6 +5632,45 @@ type Nonce struct {
 	ValidateNonce *bool `json:"validateNonce,omitempty"`
 }
 
+// OneDeployRequest - OneDeploy settings defined by user
+type OneDeployRequest struct {
+	// Specifies whether the deployment should be performed asynchronously
+	Async *bool `json:"async,omitempty"`
+
+	// Specifies whether to clean the target deployment directory
+	Clean *bool `json:"clean,omitempty"`
+
+	// Disables any language-specific defaults
+	IgnoreStack *bool `json:"ignoreStack,omitempty"`
+
+	// Kind of resource.
+	Kind *string `json:"kind,omitempty"`
+
+	// The Uri where the source artifact can be pulled from
+	PackageURI *string `json:"packageUri,omitempty"`
+
+	// The absolute path to deploy the artifact to
+	Path *string `json:"path,omitempty"`
+
+	// Resets Java apps to the default parking page if set to true with no type specified
+	Reset *bool `json:"reset,omitempty"`
+
+	// Specifies whether to restart the app following the deployment
+	Restart *bool `json:"restart,omitempty"`
+
+	// The type of the artifact being deployed
+	TrackDeploymentProgress *bool `json:"trackDeploymentProgress,omitempty"`
+
+	// READ-ONLY; Resource Id.
+	ID *string `json:"id,omitempty" azure:"ro"`
+
+	// READ-ONLY; Resource Name.
+	Name *string `json:"name,omitempty" azure:"ro"`
+
+	// READ-ONLY; Resource type.
+	Type *string `json:"type,omitempty" azure:"ro"`
+}
+
 // OpenAuthenticationAccessPolicies - AuthenticationPolicy of type Open.
 type OpenAuthenticationAccessPolicies struct {
 	// Open authentication policies.
@@ -12123,7 +12162,8 @@ type WebAppsClientCreateDeploymentSlotOptions struct {
 // WebAppsClientCreateOneDeployOperationOptions contains the optional parameters for the WebAppsClient.CreateOneDeployOperation
 // method.
 type WebAppsClientCreateOneDeployOperationOptions struct {
-	// placeholder for future optional parameters
+	// Information on OneDeploy request
+	Request *OneDeployRequest
 }
 
 // WebAppsClientCreateOrUpdateConfigurationOptions contains the optional parameters for the WebAppsClient.CreateOrUpdateConfiguration